mqtt-serial/requirements.txt
viam-sdk
robust-serial
pypubsub